Modify README and JavaScript
Modified the README.md and added some pseudo code for the function that creates our 16x16 grid.
Modified script.js to add a nested for loop that iterates 16 times for each time the outer loop is run. This now gives us 256 divs that we will need to style with flexbox.