Modify README, script, style
Modified README to make the pseudo code align more closely to what we want to happen in our loops.
Modified style.css to make the "container" have a flex display with flex-direction column, added styling for the "row-container" class as well to add flex-direction row.
Modified script.js to add a rowContainer div every outside loop iteration and append that to the "container" div. Changed the nested loop to append new divs to the rowContainer. Now, the grid is completely setup. It just needs styling.