Modified HTML, JS, and css
Modified index.html to add header with buttons and a footer.
Modified style.css to make sure things look nice. Added more containers so that the header and footer look good. Added a header that is positioned to the left of the navbar and buttons that are positioned to the right.