Modify CSS and HTML
Modified index.html to add a generate pattern button which will add a random pattern with certain blocks of the grid filled in. Functionality will be added later. Also added Roboto because it is the best font.
Modified style.css to add the Roboto font family to all text. Made the navbar and footer look better.