Modify HTML, CSS, js
Modified index.html to add two input buttons so that users can pick background and brush color.
Modified script.js so that users can use the input buttons. There are some bugs with this like the background color picker erasing all of the users strokes but I will come back to that in the future.
Generate Pattern does not work at the moment but I like the idea so I will come back to it eventually.